+++ Amplitude/AMPIdentify.h
@@ -14,6 +14,7 @@
 - (AMPIdentify*)add:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)append:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)clearAll;
+- (AMPIdentify*)prepend:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)set:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)setOnce:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)unset:(NSString*) property;