+++ Amplitude/AMPIdentify.m
@@ -69,6 +69,12 @@
     return self;
 }
 
+- (AMPIdentify*)prepend:(NSString*) property value:(NSObject*) value
+{
+    [self addToUserProperties:AMP_OP_PREPEND property:property value:value];
+    return self;
+}
+
 - (AMPIdentify*)set:(NSString*) property value:(NSObject*) value
 {
     [self addToUserProperties:AMP_OP_SET property:property value:value];